refactor: define extensible interfaces for logging & observability
Move data types (LogEntry, LogQueryOptions, SessionReportCard,
Retrospective) to types.ts as canonical contracts. Add pluggable
service interfaces:

- EventLogger: swap JSONL writer for database/cloud logging
- EventLogReader: swap JSONL reader for custom query backend
- RetrospectiveStore: swap JSON file storage for database/API

LogWriter now implements EventLogger. LifecycleManagerDeps accepts
EventLogger interface instead of hardcoded logDir string.
JsonlRetrospectiveStore provides default file-based implementation.
